adding support for OPEN and small refactor
